requirements.txt
streamlit>=1.40.1
requests>=2.32.3
beautifulsoup4>=4.12.3
markdownify>=0.14.1
openai>=1.57.1
python-dotenv>=1.0.1
pyperclip>=1.9.0 